libs: use a more configurable logging strategy
This is a big re-factoring which update all the logging statements to make
use of a more sane and consistent configuration:

- each module uses the same logger configuration but refers to a custom
  logger which has the same name of the module, this should allows a
  fine grained configuration of the loglevels for each module.

- a new libs/utils/conf.py::LisaLogging class allows to load the
  configuration and customize the root logger level.

- loggers are configured using a single logfile (logging.conf) which is
  provided by default in the LISA's root folder.

Each logger can be configured to specify:
- loglevel: INFO by default for all modules
- output channels: console and logfile (lisa.log by default) for all modules
- string format: the same by default for all modules

If logging.conf does not specify a specific configuration for a logger, than
the default logger configuration applies, which is the one defined by the
"root" logger.

Notebooks and LisaTest are required to call at least one time the
LisaLogging::setup() method to properly intialize logging.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>