tests/eas: Add generic Energy-Model based tests
These tests take an EnergyModel describing the platform, and a workload. The
workload is examined to find estimated "optimal" task placements, and
estimations of those task placements' energy efficiency. The workloads are run
on the target, and a trace is collected. This trace is processed to find the
task placement that was actually observed. Using the same metrics that were used
to decide on "optimal" behaviour, the observed behaviour is evaluated for energy
efficiency. If this energy efficiency is significantly worse than the estimated
"optimal" value then a failure is raised.

Excellent energy efficiency can be achieved by packing all tasks onto the least
capable CPU at the lowest clock frequency, but that's not likely to be
desirable. Energy efficiency is not the only criterion for good scheduler
behaviour, so another test is provided using RT-App's "slack" measurement to
assert that required throughput was provided to the workload.

Overall, this aims to allow the same tests to be run on any platform once it has
been described by an EnergyModel.

By making assertions about the estimated *efficiency* of the behaviour, rather
than the behaviour itself, the tests are made more flexible: a given workload
may be serviceable by two entirely different scheduling strategies where one is
only slightly more efficient than the other. In this case, the tests will
automatically allow either strategy to pass. The downside to this approach is
that when a test fails, it is not always obvious what scheduler behaviour was
expected or how the observed behaviour differed from this. It may be possible to
later add more user-visible representation of this information to aid failure
diagnoses.