LisaTest: Improve running tests with multiple iterations
Currently when you run a LisaTest with experiments_conf['iterations']!=0, the
test assertions are made for every experiment and the test is failed as soon as
any iteration fails. Only the first failure encountered is displayed.

With this patch all the results for each wload/conf pair are aggregated and
printed together.

An attribute `permitted_fail_prop` is added which can be overridden to allow a
certain proportion of iterations to fail without failing the overall test.

The problem is only solved across iterations: if multiple wloads or confs fail,
only the first wload/conf pair that was found to have failed will be
displayed.