utils/trace: return a single task name for each PID
On a given trace, the same PID can have different task names, mainly
because once a task is generated it inherits the parent name and then
its name is updated to represent what the task really is.

The name the task has initially is rather confusing, especially when it
comes to to a reverse mapping. In that case we do not want to report
PIDs for tasks which had a certain name only for a limited time once
initially created.

This patch works under the assumption that a task name is updated at
most one time and it always reports the name the task had the last time
it has been scheduled for execution in the current trace.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>