tests/lisa/test_trace: Factor out helper for dummy sched_switch events
We want to be able to generate made-up trace snippets for tests, which is pretty
awkward and the resulting code is really wide. This attempts to hide the
akwardness in a helper method. This approach will be used later to create a
richer trace snippet, where hopefully the trade-off makes more sense.